+++ packages/stage-ui/src/components/Scenes/VRM.vue
@@ -1,34 +1,68 @@
 <script setup lang="ts">
-import { OrbitControls } from '@tresjs/cientos'
 import { TresCanvas } from '@tresjs/core'
+import { useElementBounding } from '@vueuse/core'
 import { storeToRefs } from 'pinia'
-import { ref } from 'vue'
+import { onMounted, ref, shallowRef, watch } from 'vue'
+
+import * as THREE from 'three'
 
 import { useVRM } from '../../stores'
-import { Screen } from '../Misc'
-import { VRMModel } from '../Scenes'
-
-const props = withDefaults(defineProps<{
-  paused?: boolean
-  showAxes?: boolean
-}>(), {
-  paused: false,
-  showAxes: false,
-})
+import { OrbitControls, VRMModel } from '../Scenes'
 
 const emit = defineEmits<{
   (e: 'loadModelProgress', value: number): void
   (e: 'error', value: unknown): void
 }>()
-const { selectedModel, scale } = storeToRefs(useVRM())
+const VRMContainerRef = ref<HTMLDivElement>()
+const { width, height } = useElementBounding(VRMContainerRef)
+const {
+  selectedModel,
+  cameraFOV,
+  initialCameraPosition,
+} = storeToRefs(useVRM())
 
-const cameraPositionX = ref(0.17)
-const cameraPositionY = ref(0.5)
-const cameraPositionZ = ref(-1)
+const modelRef = ref<InstanceType<typeof VRMModel>>()
 
-const modelRef = ref<{
-  setExpression: (expression: string) => void
-}>()
+const camera = shallowRef(new THREE.PerspectiveCamera())
+
+onMounted(() => {
+  if (VRMContainerRef.value) {
+    camera.value.aspect = width.value / height.value
+    camera.value.fov = cameraFOV.value
+    camera.value.position.set(
+      initialCameraPosition.value.x,
+      initialCameraPosition.value.y,
+      initialCameraPosition.value.z,
+    )
+    camera.value.updateProjectionMatrix()
+  }
+})
+
+watch(cameraFOV, (newFov) => {
+  if (camera.value) {
+    camera.value.fov = newFov
+    camera.value.updateProjectionMatrix()
+  }
+})
+
+// not sure if we need this
+// watch([width, height], () => {
+//   if (camera.value) {
+//     camera.value.aspect = width.value / height.value
+//     camera.value.updateProjectionMatrix()
+//   }
+// })
+
+watch(selectedModel, () => {
+  if (camera.value) {
+    camera.value.position.set(
+      initialCameraPosition.value.x,
+      initialCameraPosition.value.y,
+      initialCameraPosition.value.z,
+    )
+    camera.value.updateProjectionMatrix()
+  }
+})
 
 defineExpose({
   setExpression: (expression: string) => {
@@ -38,11 +72,9 @@ defineExpose({
 </script>
 
 <template>
-  <Screen v-slot="{ width, height }" relative>
-    <TresCanvas :alpha="true" :antialias="true" :width="width" :height="height">
-      <OrbitControls :enable-zoom="false" />
-      <TresAxesHelper v-if="props.showAxes" :size="1" />
-      <TresPerspectiveCamera :position="[cameraPositionX, cameraPositionY, cameraPositionZ]" :zoom="scale" />
+  <div ref="VRMContainerRef" w="100%" h="100%">
+    <TresCanvas v-if="camera" :camera="camera" :alpha="true" :antialias="true" :width="width" :height="height">
+      <TresAxesHelper :size="1" />
       <TresDirectionalLight :color="0xFFFFFF" :intensity="1.2" :position="[1, 1, 1]" />
       <TresAmbientLight :color="0xFFFFFF" :intensity="1.5" />
       <VRMModel
@@ -50,10 +82,12 @@ defineExpose({
         :key="selectedModel"
         :model="selectedModel"
         idle-animation="/assets/vrm/animations/idle_loop.vrma"
-        :paused="props.paused"
+        :paused="false"
         @load-model-progress="(val) => emit('loadModelProgress', val)"
         @error="(val) => emit('error', val)"
       />
+      <!-- <TresPerspectiveCamera :position="[initialCameraPosition.x, initialCameraPosition.y, initialCameraPosition.z]" :fov="cameraFOV"/> -->
+      <OrbitControls />
     </TresCanvas>
-  </Screen>
+  </div>
 </template>
